refactor: CurtomError.js 파일을 modesl-> custromError 폴더로 옮기고 파일명을 index.js로 변경
