+++ src/models/index.js
@@ -1,7 +1,5 @@
 import Badge from './Badge.js';
 
-import CustomError from './CustomError.js';
-
 import {
   GiftEvent,
   SpecialEvent,
@@ -18,7 +16,6 @@ import ReservationDate from './ReservationDate.js';
 
 export {
   Badge,
-  CustomError,
   Order,
   OrderedMenu,
   GiftEvent,
